+++ src/Pages/Dashboard/MyOrders/MyOrders.js
@@ -14,19 +14,21 @@ const MyOrders = () => {
                 setMyOrders(data)
                 console.log(data);
             });
-    }, [user?.email,control]);
+    }, [user?.email, control]);
 
     const handleDelete = (id) => {
-        fetch(`http://localhost:5000/deleteOrder/${id}`, {
-            method: "DELETE",
-        })
-            .then((res) => res.json())
-            .then((data) => {
-                if (data.deletedCount) {
-                    setControl(!control);
-                }
-            });
-        console.log(id);
+        const proceed = window.confirm('Are you sure you want to delete this ?')
+        if (proceed) {
+            fetch(`http://localhost:5000/deleteOrder/${id}`, {
+                method: "DELETE",
+            })
+                .then((res) => res.json())
+                .then((data) => {
+                    if (data.deletedCount) {
+                        setControl(!control);
+                    }
+                });
+        }
     };
 
     return (
@@ -50,7 +52,7 @@ const MyOrders = () => {
                     {
                         myOrders?.map((pd, index) => (
                             <tbody
-                            key={pd._id}
+                                key={pd._id}
                             >
                                 <tr>
                                     <td data-label="S.No">{index}</td>
@@ -60,8 +62,8 @@ const MyOrders = () => {
                                     <td data-label="Status">{pd?.status}</td>
                                     <td className="table_btn" data-label="Cancel">
                                         <button
-                                        onClick={() => handleDelete(pd?._id)}
-                                         className="btn">Cancel</button>
+                                            onClick={() => handleDelete(pd?._id)}
+                                            className="btn">Cancel</button>
                                     </td>
                                 </tr>
